Fixes and Cleanups for EF Code First setup
Work towards closing #14.

Removed IDbSets for Address, CartItem, MemberCreditCard, OrderItem, and
Person.

Removed the EF conventions to pluralize tables and cascade delete on one
to many relationships.

Added NotForSale GameStatus.
Renamed Discontinued GameStatus to DiscontinuedByManufacturer.

Removed the Address class due to issues with composite keys from base
classes in Entity Framework. As a result, all of the fields were moved
into Location and MemberAddress.

Renamed CreditCardPaymentInformation to MemberCreditCard.
Created CreditCardBillingInfo to store address information for credit
cards. This is mapped to the same table as MemberCreditCard in EF.
Added CreditCardBillingInfo property to MemberCreditCard.

Fixed key types in ProductLocationInventory.

Added PublishedGameProducts and DevelopedGameProducts to Company instead
of having just GameProducts as EF understandably doesn't like two one to
many relationships joining on the same property. It is however fine with
you doing that as long as you define the relationships on the side with
many.

Added a package which should let us easily map enums to lookup tables.
This hasn't been tested or implemented yet.