Setup ASP.NET Identity
More work towards #14
Added Identity classes with GUID keys.
Changed Identity infrastructure to use GUID keys.
Moved Identity classes from Veil to Veil.DataModels. This has the unfortunate side effect of the project now having a dependency on  Microsoft.AspNet.Identity.Core and Microsoft.AspNet.Identity.EntityFramework

Added IIdentityExtensions class with an extension method for getting the User's Id as the default infrastructure only supports string Ids. We may want to add checking where we use this to ensure we didn't get an empty GUID back.

Renamed ApplicationUser to User.
Moved the FirstName and LastName fields from Person to User. User already contains Email and PhoneNumber.
Removed the Person class.
Added navigation property to User in Member and Employee.

Fixed issue where Unity was trying to resolve the dependencies for the AccountController but was unable to do so as we haven't registered the dependencies.

Updated EF config to autogenerate keys as it doesn't do that by default.
Changed base class for VeilDataContext to IdentityDbContext so that we don't need to DbContext classes.