Fixed Member to Cart relationship.
I changed the Member to Cart relationship to be required:required. This allows us to map the Cart to the Member table. As the cart has no extra properties, this adds nothing to the Member table. However, removes the awkward Cart table which only had a single column (MemberId) as a PK and FK.

The downside to this is that if we ever add properties to the cart in the future we will either have to accept them being added to the Member table or change Cart to be its own table and not mapped to the Member table.