Primarily DB related changes.
Changed non account-related controllers to use DI.
Removed Dispose methods from the controllers as Unity should take care of this for us.

Changed max length of ESRBRatingId to 8 from 20.

Moved EnumToLookup from the constructor of VeilDataContext to the Migration Seed method.
Added Seed values for ESRB Ratings and ESRB Content Descriptors.

Added missing links to the Reports index page.

Fixed the top link on the billing information checkout page which was linking back to the same page.

Added comments to IVeilDataAccess.
Added generic MarkAsModified to IVeilDataAccess and VeilDataContext.
Added SaveChangesAsync to IVeilDataAccess.

Biggest lesson learned from this commit: If googling doesn't show a bunch of people (anyone) with the same issue as you, it is probably that very uncommon NuGet package (EnumToLookup) you are using that is causing the problem. A couple hours wasted on that one!