Added tests for VeilDataContext.
These tests are currently set up to run on SQL Server 2014's version of LocalDB which has a name of MSSQLLocalDB.
It currently uses the 'master' System Database which isn't the greatest but required the least amount of work to do.
Both the test classes setup and tear down the database via migrations, except for the EnumToLookup tables which are left behind as that isn't part of our migrations.

I haven't added tests for CRUD actions as I think we can safely assume the EF team has their own tests for this and we don't have any customization of these methods. However, we may want to look at having integration tests between the create/update controller actions and the database as it will any expose validation errors we missed.

I exposed the internal DataAccess.Migrations.Configuration class to DataAccess.Tests to allow for the included migration tests.

Upgraded NuGet Packages.
Removed unused EF NuGet reference in the Models project.