Register/Login Work
Removed Site.css which contained bootstrap override styles which were
breaking things.
Removed ReadMe for foundation.

Added friendly display name to WishListVisibility.FriendsOnly.

## Account Controller

Removed Get Register action. As we are going with a combined
Login/Register page this isn't needed.
Moved all of the currently unused methods to near the bottom of the
controller in a region.
Implemented Member and StripeCustomer creation as part of the Register
POST action.
Removed the Dispose override as Unity handles that for us.

## Home Controller

Removed the Stripe examples as they now exist in StripeService.

## Manage Controller

Broke ManageAddresses and ManageCreditCards out into their own
actions/views and set up the Country/Province list part of them.

## Helpers

Added extension methods for creating select lists for Countries and
Provinces from IEnumerable<Country>.

## View Models

Added LoginRegisterViewModel for the Account/Login view. It simply has
properties for a LoginViewModel and a RegisterViewModel.
Added ReceivePromotionalEmail and WishListVisibility properties to
RegisterViewModel.

Added ManageAddressViewModel for the ManageAddresses view.
Added ManageCreditCardViewModel for the ManageCreditCards view.

## sass

Added classes to make it so we can put checkboxes and labels on a single
line side by side.
Added a class to add some top padding to elements. This could maybe be
done via Foundation? I'm not sure.

## Stripe Service

Added StripeService and IStripeService.
Setup Unity to resolve IStripeService.

## Views
### Account

Changed the Login view to use _LoginAccountPartial and
_RegisterAccountPartial.
Changed Login view to have LoginRegisterViewModel as its model.
Removed Register view.
Setup the Login view to only display validation errors on the form which
had validation errors.

### Manage

Added ManageAddress and ManageCreditCards instead of having them as part
of Index.

### Shared
Renamed _LoginPartial to _LoginNavBarPartial to avoid any chance of
confusion.