Work towards #29 #30 Register/Login
#DataAccess

## IVeilDataAccess
Added an IUserStore property to the interface to make it clearer where the UserStore used by VeilUserManager is coming from.

## Migrations.Configuration
Added seed values for our 3 primary roles (Admin, Employee, Member)

## VeilDataContext
Added the new UserStore property required as part of IVeilDataAccess

# Models

## VeilRoles
Added this static class with property for getting the role names for our app so that we don't have those strings duplicated all over the place.

# Veil

## Startup.Auth.cs
Reduced validateInterval from 30 minutes to 10 minutes.

## UnityConfig
Removed registering VeilDataContext as the source for DbContext injections.
Removed registration of IUserStore<User, Guid> as this is now supplied via IVeilDataAccess. I made this change because it was very difficult to follow the construction chain for it. It was making a UserStore which requires a DbContext as a constructor argument. This DbContext was then resolved by Unity as VeilDataContext.

## Content
Removed the compiled foundation css files and added them to the .gitignore

## AccountController
Added IVeilDataAccess parameter to the ctor.

Added functionality to confirm a user is in the correct roles when they log in. This is implemented in EnsureCorrectRolesAsync.
If the user's roles are change, we sign them out and then immediately sign them back in. This is the best work around I could find for immediately propagating the role changes.

Added back the GET action for Register. This is required because if the POST action for Register fails it doesn't redirect to Login but rather just shows the login view.

Added a transaction to Register to ensure we don't create a User and their associated Member unless we successfully create a Stripe customer for them.

## ManageController
Converted the chained ternary for statusMessage in Index to a switch.

Changed status message from being set on ViewBag to being set on the IndexViewModel.

Added RememberBrowser and ForgetBrowser actions for two factor auth.

Added helper method for getting the current users Id as a Guid.

## WebOrdersController
Added the actions for unprocessed orders to this controller.

## IIdentityExtensions
Changed visibility from internal to public.

## ManageViewModels
Added ReceivePromotionalEmail property.
Added StatusMessage property. This used to be set in ViewBag instead.

## StripeService
Added some initial documentation to CreateCustomer.

## VeilUserManager
Changed constructor from taking IUserStore to taking IVeilDataAccess and then using its UserStore property.

Changed password validation logic to be less strict in debug mode

## Manage/Index
Get status message from the Model instead of ViewBag.
Added the missing Forget/Remember browser buttons for two factor auth.

## Manage/ManageCreditCards
Corrected the property for the address to be creditCard.AddressState

## Shared/_LoginNavBarPartial
Implemented role checking for the My Stuff dropdown so only the appropriate actions are shown. Currently this is only enforced when Web.Config ```<compilation debug="true" ..```. is set.

## Shared/_NavBarPartial
Removed the My Stuff dropdown.

## WebOrders
Added blank views for UnprocessedOrderDetails and UnprocessedOrders