Account Login/Register Work #29 #30
## Account Controller
Added documentation comments to all the actions and methods I consider implementation complete.

Cleaned up EnsureCorrectRolesAsync now that we needed the User to get to the point we are calling it.

Extracted the code for sending a email confirmation email into the SendConfirmationEmail method.

Added invalidation of all tokens after the user successfully confirms their email. This is mainly to invalidate the email link.
Added basic error handling to the ConfirmEmail action.

Added actions for resending confirmation emails. This is shown if a user attempts to log in and their credentials are valid but their email is unconfirmed.

Added support for the returnUrl to the register actions. This will only ever be used if the user tries to sign up, it fails, and then the user logs with a valid account.

Fixed the first parameter to PasswordSingInAsync. It wants username but we were giving it email. This worked before now because we were incorrectly setting username as the email despite the fact that we have a username field.

## AccountViewModels
Added missing validation attributes that were on User but not RegisterViewModel

## EmailService
Implemented IIdentityMessageService using SendGrid.
I've wrapped the call in a try catch but currently I am just writing the exception message via Debug.WriteLine.

## VeilUserManager
Set the token lifespan for (all?) tokens created to 12 hours.

## Views
### ConfirmEmail
Updated the styling to match our other pages.

### ConfirmResendConfirmationEmail
Added.
This view lets the user resend confirmation emails.

### RegisterComplete
Added.
This view informs the user that they must confirm their email

### ResendConfirmationEmail
Added.
This is the view that is shown after the user clicks the resend confirmation email on the ConfirmResendConfirmationEmail view.

### _RegisterAccountPartial
Added return url to the form postback url.

### Shared/Error
Added a ValidationSummary. This is currently only used by AccountController.ConfirmEmail