Model & Account Controller Changes
## User
Added/fixed doc comments

Added Email override property to add validation attributes to it
Added MaxLength property to UserName. This is the length of 256 as IdentityDbContext's OnModelCreating. I don't think it is worth the tiny amount of effort required to override this to be 255 like all other columns of this length.

## PhysicalGameProduct
Reduced MaxLength to 255 on SKUNameSuffix from 256.

## Product
Added MaxLength of 2048 on BoxArtImageURL.

## WebOrder
Made the ProcessedDate nullable. I missed this when originally creating it.

## Account Controller
Added redirects to the home page when authenticated users go to /Login, /Register, /ResendConfirmationEmail.

## AccountViewModels
Added MaxLength of 255 to Username on RegisterViewModel

## EmailService
Inlined the SendEmail method directly into SendAsync

## _LoginNavBarPartial
Changed order of Reports and Unprocessed Orders to place Unprocessed Orders first.

## _NavBarPartial
Changed cart to only render if the User is a Member. This still always shows as there is a IsDebuggingEnabled check too.

## Migrations
Redid Migrations to integrate the model changes.