Changed VeilRoles from properties to constants
This allows them to be used in attributes and other locations which require constants.

## ManageController
Added a fix for something which will only really happen in development. If you signed in, deleted that user, and then try to view any action which attempts to find the user by their id, you will get an exception because it is unable to find a match.