Game.GameAvailabilityStatus Changes/Fixes
## Game
GameAvailabilityStatus is now a computed column.
As noted in the doc comments, you must explicitly reload the game if you modify the GameSKUs in a way that would change this computed value. I have an idea of how to fixed this but am not implementing it in this commit.

## AvailabilityStatus Enum
Added explicit values to the enum values as they are used in a stored function which computes a Game's AvailabilityStatus.

## GetGameAvailabilityStatusTests
This file constains integration tests to ensure the function for computing a Game's Availability status is correct.

## VeilDataAccess.Configuration.Seed
Renamed Personal Computer to PC
Added some seed tag values for the seed games.

Commented out the EnumToLookup code as it is now a migration.

## Migrations
Added a migration to create the function which computes a Game's AvailabilityStatus

Added a migration to add Enum lookup tables.