Added Friendship Insert Trigger
This insert trigger ensures the lesser of the two MemberIds is always inserted as the first column.

This seems to be the easiest way to enforce uniqueness across the two combinations of columns { (RequesterId, ReceiverId), (ReceiverId, RequesterId) }.

It does, however, make the column names inaccurate. I don't know if we care about that or not.