Adding unified 404 error model
Added base controller which handles HttpExceptions with a status code of 404.

## GamesController
Set up Details to use the new 404 error handling.

Made all controllere except Account inherit from this.

## CommonErrorResults
This class contains Controller extensions methods for common error results like Not Found.

## ErrorViewModel
Added this as a model for the error pages

## Views\Shares\NotFound
Common view for 404 Not Found