Games Controller work
Added some missing doc comments

## Search
Updated to support the new paginated Index view.
Updated tests to pass with the new view model.

## Details
Changed this back to using Game as a model.
Removed GameDetailsViewModel.

## RenderPhysicalGameProductPartial
Moved the current member portion of the removed GameDetailsViewModel into RenderPhysicalGameProductPartial. This allows Details and this partial to be more loosely coupled.

## _PhysicalGameProductPartial
Fixed the link for deleting the physical game product