GamesController Work #26, #27, #28
## GamesController
Added GamesPerPage property to allow a seam for testing pagination.

Moved HTTP_NOT_FOUND to BaseController and renamed it NotFound.

Added helper method for filtering out not for sale games for non-employee/admin users.

Changed Search to not bother filtering by the keyword when it is empty.

## _NavBarPartial
Changed the search to be a GET instead of POST.

## GamesControllerTests
Added more tests for Search:
- Ensure we don't filter with an empty keyword.
- Ensure we order by name
- Ensure we are setting current page to the passed in value
- Ensure total pages count is set and accurate.
- Ensure we are limiting to the GamesPerPage value
- Ensure we return an empty list if someone goes to a page which is past where we would have enough games.

## GameListViewModelTests
Added tests for CurrentPage now that it is more than a simple autoproperty.