GamesController Work #22, #25
## GamesController
### DeleteGameConfirmed
Fixed this to work. It was trying to delete items inside a loop on the collection.
Made this catch the appropriate exception for FK violations.

### DeleteGameProductConfirmed
Made this catch the appropriate exception for FK violations.

### CreatePhysicalGameProduct
Fixed the prefixes for internal SKUs.

## DataAccess
Added migration to remove Many To Many Casecade Delete by default. We will want to go through all of these and decide which ones make sense.

## Models
Moved NewInventory and UsedInventory up to Product from PhysicalGameProduct. Althought DownloadGameProduct doesn't need these properties, inventory levels are stored for Products so it fits better at that level.

## Views\Games\_PhysicalGameProductPartial
Fixed the Edit/Delete links.