Tests for Add Shipping Address (#43)
# IVeilDataAccess
Added an enum for SQL Error numbers. Currently it only has the number for a constraint violation.

# ManageController
Cleaned up the handling of the foreign key constraint violation.

# ManageControllerTests
Added Tests for ManageAddresses and CreateAddress.

# SqlExceptionCreator
Added this so that I could test the handling of a foreign key constraint violation. SqlException is sealed so you need to use reflection to create one and set it up.

# TestHelpers
Added methods which help setup when testing methods which get the current Users Id.

- A method for getting a controller context which is set up to return a value when User.Identity is called
- A method for getting a faked IGuidUserIdGetter which returns the given Guid.