Tests for Add Shipping Address #43
#Models\MemberAddress
Added RegularExpressionAttribute for the store postal code formats we allow.

#Models\ValidationRegex
Added regexes for Input and stored Postal/Zip codes

#Models.Tests\Validation\ValidationRegexTests
Added tests for the new postal code regexes
Updated existing tests to use Is.StringMatching instead of calling Regex.IsMatch

#Veil\Controllers\ManageController
Added formating of the input postal/zip code into the format we want to store them in. This is implemented in a way that doesn't allow for more countries to be added, but I don't consider it to be worthwhile to do the small amount of work to required add a PostalCodeRegex column to the Country table. Realistically, we would likely not even be selling outside of Canada, let alone Canada and the US.

Also made it so the validation message for the postal code will be more specific if possible.

#Veil\Models\ManageAddressViewModel
Added RegularExpression attribute for postal code.

#Veil.Tests\Controlls\ManageControllerTests
Added tests for invalid postal code model state.
Added tests for the standardized postal code format being saved to the database.