Implementation for (#44) Add Billing Information
# Controllers
## AccountController
Made the two model error key public so the views could use them.

##ManageController
Implemented adding credit cards.
With this implementation, we never have access to the card information so we have don't have to meet any PCI security requirements.

Renamed SetupCountriesAndAddresses to SetupAddressesAndCountries.

# Extensions\HtmlHelperExtensions
Added Html helpers for creating dropdowns and editors without the name attribute. This was required because we don't want any of the credit card information being sent to stripe to also be posted back to us.

# Models
## CreditCardViewModel
Removed StripeCustomerId
Added display names and other missing attributes.

## ManageCreditCardViewModel
Added SelectListItem properties for years, months, and the member's credit cards.
Fixed CreditCard not being a property.

# sass\Site.scss
Added classes for the stripe error part on ManageCreditCards.

Added class for validation summaries. I'm not sure if this is actually used anywhere, but I did the work before deciding validation summary didn't work for my situation on ManageCreditCards.

# Services
## Interfaces\IStripeService
Update CreateCreditCard to take a card token instead of a CreditCardViewModel.
Updated CreateCreditCard to return a MemberCreditCard instead of a string.

## StripeService
Fully implemented CreateCreditCard.

#Views
##Account
Updated the Login and Register partials to use the model error key constants from AccountController instead of strings.

## Manage
### ManageAddresses
Fixed issue with a check in the javascript to make sure we didn't override the preselected province/state on page load.

### ManageCreditCards
Fully implemented the Add part of this view. The delete portion is almost completely implemented but a controller action for it doesn't exist.

### _ManageCreditCardFormFields
This contains the form fields for the ManageCreditCardViewModel so that this partial can be reused for checkout.