Tests for Add Billing Information (#44)
# ManageController
Fixed the call to AddModelError not using the Stripe Issues key constant.
Changed the error path to simply redirect to ManageCreditCards as we don't have any information to redisplay anyways.

# CreditCardViewModel
Made this abstract as its only function is to help when generating views.

# ManageCreditCardViewModel
Added doc comments.
Changed CreditCard property to be get only as it is never meant to be assigned anything.