Shipping Info Checkout Step for #48
This commit contains significant changes to the database in terms of a
web order's address.
In addition, the primary key for CartItem has been
changed from { MemberId, ProductId } to  { MemberId, ProductId, and
IsNew }.
Finally, WebOrder now has no link to MemberCreditCard. It
simply keeps the StripeCharge token and the last 4 digits of the card
used for the order.

# DataAccess
## EntityConfigurations
###
CartItemEntityConfig
Updated primary key to include IsNew

###
MemberAddressEntityConfig
Updated this to reflect the fact that a subset
of the address information is now in the Address complex type.

### WebOrderEntityConfig
Removed MemberAddress setup.
Added setup for address information.
Removed MemberCreditCard setup.

## Migrations
Added migration to which adds address information to WebOrder.
Added migration to remove the FK for MemberCreditCard and add the credit
card's last 4 digits instead.
Added migration to include IsNew as part of the primary key for CartItem

Added a default Location to the seed method.

# Models\Models
## Address
This is a complex type which is mapped directly into the table for
entities which contain it.
Unfortunately, Province/Country and their codes can't be part of this
type if we want to have foreign keys and enforce FK integrity.

## CartItem
Added Key attribute to IsNew.

## MemberAddress
Added Address property which contains the other properties which were
removed.

## WebOrder
Added address information properties.
Removed MemberCreditCard properties.
Added CreditCardLast4Digits property.

# Veil
## Controllers
### CheckoutController
Implemented the ShippingInfo portion.

### ManageController
Moved the method for setting up AddressViewModel onto the view model so
it can be reused by multiple controllers.

Moved the logic for making the PostalCode model error message more
specific to AddressViewModel so that it can be reused by multiple
controllers.

Moved the logic for converting PostalCode into the persisted format into
AddressViewModel so that it can be reused by multiple controllers.

### WebOrdersController
Removed the Includes of properties that are no longer on WebOrder.

## Models
Renamed ManageAddressviewModel toAddressViewModel and added common logic
used by the Manage and Checkout controllers.

Renamed ManageCreditCardViewModel to BillingInfoViewModel and this will
contain common logic used by the Manage and Checkout controllers.

Added WebOrderCheckoutDetails which is stored in the Session as the user
checks out.

## sass\Site.scss
Added comments to some styles.

## Views
### Checkout\BillingInfo
Very minor initial work on implementing this view.

### Checkout\ShippingInfo
Fully implemented this view.

### Manage\ManageAddresses
Changed view model name

### Shared\NotFound
Removed ValidationSummary.

### Shared\AddressViewModel\_AddressFormFields
Moved and renamed this from Manage views