Work toward Checkout #48
Mainly work towards the Confirm order part of checkout. This also meant I realized we weren't doing anythng with Inventory anywhere else so added that in the few places I thought it should be.

# DataAccess\Migrations\Configuration
Added seed values for inventory levels of our seed games.

# Models\Models
## CartItem
Implemented an IEqualityComparer so that I can check if the POSTed cart is the same as the one in the database.

## Location
Added a constant for the Online warehouse name. The idea is to use this for all inventory related things in our system.

# Veil\Controllers
## CheckoutController
Lots of refactoring to break the logic in ConfirmOrder and PlaceOrder out into separate methods.

Changed PlaceOrder to take back the list of items in the cart as it was when the page was displayed. This allows us to confirm the cart hasn't changed and thus don't charge the user more/less than they agreed to.

Added the logic for decreasing inventory levels. This currently throws an InvalidOperationException if we are in a state where decreasing the inventory level isn't valid.

## GameProductsController
Added removal of the 0 stock ProductLocationInventories to the DeleteConfirmed action. If inventories exist with levels > 0, this will result in a foreign key exception.

Added adding a ProductLocationInventory entry for the online warehouse when creating a PhysicalGameSKU.

Both of these changes are effectively untested. I only updated the tests to continue passing and nothing more. I've placed TODO:s around the new code.

## GamesController
Added removal of the 0 stock ProductLocationInventories to the DeleteGameConfirmed action. If inventories exist with levels > 0, this will result in a foreign key exception.

This code is also effectively untested. This required no test changes as there weren't any with ProductLocationInventories for the Game's SKUs. I've placed TODO:s around the new code.

# Veil\Models\ConfirmOrderViewModel
Changed IsNew from a string to a bool.

# Veil\Views\Checkout\ConfirmOrder
Add hidden inputs for all the items in the cart so we can confirm cart equivalence.

# Veil.Tests
## Controllers\GameProductsControllerTests
Updated the tests that failed due to me adding code related to ProductLocationInventories to pass and nothing more.

## TestHelpers
Added helper method for getting a Mock<DbSet<Location>> which contains the online warehouse location.