Work on #48
Added a custom exception for Not Enough Inventory and added a catch for it.

Reset session cart quantity on order completetion.

Added link to the order detail in the order success alert.

Updated alert helper to accept a link. This is rendered using Html.Raw so it should NOT contain ANY user input.
Added styles for links in alert boxes.

Moved My Events from _NavBarPartial to _LoginNavBarPartial.