Cascade Delete changes
Added cascade delete for both GameProduct and Member FKs of GameReview.

Added cascade delete for Member FK of MemberCreditCard.

Added cascade delete for the User FK of UserLogin and UserClaim.

Cleaned up comments and lamba variable names.

Modified the method for getting the current cart count to only retrieve the cart count rather than pulling in the member.

Reduced nesting of GameProductsController DeleteConfirmed.

Added product property to NotEnoughInventoryException as it would probably be useful for logging purposes.

Fixed styling of the Error shared page.

Added unimplemented test for DeleteGameProduct which will need to be implemented. This action currently fails when the SaveChangesAsync call throws. It fails because the model's Game and Platform properties are nulled out by EF but the view is expecting them.