Cleanup/Doc work on #48 Checkout
# CheckoutController
Moved the ChargeCard call to after we have confirmed inventory levels are okay.

Added doc comments to everything.
Cleaned up the code as best as I could.

# IPrincipalExtensions
Finally fixed the spelling issue.

# StringExtensions
Added an extension method for formatting the last 4 digits of a card.

# ConfirmOrderViewModel
Renamed ConfirmOrderCartItem to ConfirmOrderCartItemViewModel to make it clearer it is a view model.

# IStripeService
Moved doc comments from the implementation to the interface.

# StripeService
Moved doc comments from the implementation to the interface.
Added optional Description argument to ChargeCard

# _OrderConfirmationEmail
Moved this to the correct view folder.

# WebOrders\Details
Changed this to use the new FormatLast4Digits method.