Checkout Tests skeletons #48
Moved JavaScript library files into Scripts\Library

Cleanup/reorder stuff in CheckoutController as a result of going through line by line for tests.

Moved the javascript for adding a Stripe card to AddStripeCard. Anything using this must give the form an Id of addCardForm

Fixed issue with the Stripe add card script removing the name attribute of the saveCard checkbox on BillingInfo.