Home/GameDetail small-screen fixes
Foundation small-* is bad.

Changed the video embed on Games\Details to use iframe instead of embed as embed doesn't work in firefox.

Added validation for the Game.TrailerURL to prevent it from being anything but a youtube embed link. This is required because we are using an iframe to host the Trailer and don't want to allow anything nefarious to be used.